requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.2
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
idna==3.11
imbalanced-learn==0.12.0
matplotlib==3.8.0
numpy==1.26.0
pandas==2.1.0
Pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
requests==2.32.5
scikit-learn==1.4.0
scipy==1.12.0
sqlparse==0.5.5
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.0
tzdata==2025.3
urllib3==2.6.3
xgboost==2.0.0